Switch from zope.testrunner to stock unittest
