+++ buildout.cfg
@@ -3,8 +3,7 @@ develop = .
 extends = versions.cfg
 versions = versions
 show-picked-versions = true
-parts = test
-        scripts
+parts = scripts
         crate
         omelette
         tox
@@ -14,7 +13,7 @@ parts = test
 [scripts]
 recipe = zc.recipe.egg:script
 interpreter = py
-eggs = crash
+eggs = crash [test]
        wheel
        codecov
        coverage
@@ -25,13 +24,6 @@ recipe = hexagonit.recipe.download
 url = https://cdn.crate.io/downloads/releases/crate-${versions:crate_server}.tar.gz
 strip-top-level-dir = true
 ignore-existing = true
-
-[test]
-relative-paths=true
-recipe = zc.recipe.testrunner
-working-directory = ${buildout:directory}/docs
-defaults = ['--auto-color']
-eggs = crash [test]
 
 [linter]
 recipe = zc.recipe.egg:script
