+++ tox.ini
@@ -6,7 +6,6 @@ usedevelop = True
 deps =
     crate[test] >= 0.13.2
     prompt-toolkit
-    zope.testrunner
     colorama
 commands =
-    zope-testrunner -c
+    python -m unittest -v
